Added autoSubliminalToSickbeard post processing
If you use separate folders for Auto-Subliminal and Sickbeard post
processing, you can use this script to move the subtitle and the video
file to the Sickbeard post processing folder once the subitlte is
downloaded. Afterwards it checks if it can also do a cleanup of left
over files in the Auto-Subliminal folder.