Add custom addic7ed provider
A new custom addic7ed provider is now available in order to decrease the number of 304 responses.
The provider is a clone from the subliminal version, but with a random user agent.
The random user agent should decrease the number of 304 responses of the addic7ed website (addic7ed blocks requests based on user agent).