Add option to scan for hardcoded subtitles
We are not able to check if a video contains a hardcoded subtitle, so we detect it by checking if there is a file next to the video file with the same name, but with .hardcoded.subs extension.
If this file is present, the first line in the file contains a comma separated list of the hardcoded subtitle languages.